https://github.com/javers/javers/issues/666 refactor package name, take 2.
